refactor dashboard stats architecture and restore test/runtime scope
