+++ app/Providers/AppServiceProvider.php
@@ -2,8 +2,12 @@
 
 namespace App\Providers;
 
+use App\Models\Admin\WorkDailyLog;
+use App\Models\Admin\WorkDoc;
 use App\Models\User\User;
 use App\Models\Web\Comment;
+use App\Observers\Admin\WorkDailyLogObserver;
+use App\Observers\Admin\WorkDocObserver;
 use App\Observers\UserObserver;
 use App\Observers\Web\CommentObserver;
 use Illuminate\Http\Resources\Json\JsonResource;
@@ -26,6 +30,8 @@ class AppServiceProvider extends ServiceProvider
     {
         Comment::observe(CommentObserver::class);
         User::observe(UserObserver::class);
+        WorkDailyLog::observe(WorkDailyLogObserver::class);
+        WorkDoc::observe(WorkDocObserver::class);
         JsonResource::withoutWrapping();
     }
 }
