--- tests/Pest.php
@@ -1,5 +1,0 @@
-<?php
-
-use Tests\TestCase;
-
-uses(TestCase::class)->in('Feature');
